+++ layout.py
@@ -1,0 +1,139 @@
+toppage_css = """
+    <style>
+    div.stButton {
+            text-align: center;
+    }
+    div.stButton > button {
+            text-align: center;
+            width: 500px;
+            height: 40px;
+            color: #5f6c7b;
+    }
+    /* ボタン色変更 */
+    div.stButton > button:hover {
+        border-color: #90b4ce;
+        color: #90b4ce;
+        # border-color: #3da9fc;
+        # color: #3da9fc;
+    }
+    div.stButton > button:focus:not(:active) {
+        border-color: #90b4ce;
+        color: #90b4ce;
+    }
+    div.stButton > button.css-1cpxqw2:focus {
+        box-shadow: unset;
+    }
+    div.stButton > button.css-1cpxqw2:focus {
+        border-color: #90b4ce;
+        background-color: #90b4ce;
+        color: #fffffe;
+    }
+    /* ボタン色変更ここまで */
+    </style>
+    """
+
+rem_start_button_css = """
+    <style>
+    div.css-1n76uvr > div.element-container > div.stButton > button.edgvbvh9 {
+        display: none;
+    }
+    </style>
+    """
+
+rem_stop_button_css = """
+    <style>
+    div.css-1n76uvr > div > div.css-1n76uvr > div.element-container > div.stButton > button.edgvbvh9 {
+        display:none;
+    }
+    </style>
+    """
+
+button_css = """
+    <style>
+    div.stButton {
+        text-align: center;
+    }
+    div.stButton > button {
+        text-align: center;
+        width: 91px;
+        height: 40px;
+        position: relative;
+        right: 308px;
+        bottom: -170px;
+        z-index: 61;
+        color: #5f6c7b;
+    }
+    /* ボタン色変更 */
+    div.stButton > button:hover {
+        border-color: #90b4ce;
+        color: #90b4ce;
+        # border-color: #3da9fc;
+        # color: #3da9fc;
+    }
+    div.stButton > button:focus:not(:active) {
+        border-color: #90b4ce;
+        color: #90b4ce;
+    }
+    div.stButton > button.css-1cpxqw2:focus {
+        box-shadow: unset;
+    }
+    div.stButton > button.css-1cpxqw2:focus {
+        border-color: #90b4ce;
+        background-color: #90b4ce;
+        color: #fffffe;
+    }
+    /* ボタン色変更ここまで */
+    </style>
+    """
+
+stop_css = """
+    <style>
+    div.element-container > div.stButton > button {
+        text-align: center;
+        width: 100px;
+        height: 40px;
+        position: relative;
+        right: 302px;
+        bottom: 78px;
+        z-index: 61;
+        # opacity:0;
+        # animation-name: sample01; /*←@keyframesにも同じ名前を記述*/
+        # animation-duration: 3s;
+        # animation-fill-mode: forwards;
+    }
+    </style>
+    """
+
+result_css = """
+    <style>
+    div.stButton {
+        text-align: center;
+        margin: 0 0;
+        color: #5f6c7b;
+    }
+
+    .vsc-initialized {
+        margin: 0;
+    }
+    /* ボタン色変更 */
+    div.stButton > button:hover {
+        border-color: #90b4ce;
+        color: #90b4ce;
+        # border-color: #3da9fc;
+        # color: #3da9fc;
+    }
+    div.stButton > button:focus:not(:active) {
+        border-color: #90b4ce;
+        color: #90b4ce;
+    }
+    div.stButton > button.css-1cpxqw2:focus {
+        box-shadow: unset;
+    }
+    div.stButton > button.css-1cpxqw2:focus {
+        border-color: #90b4ce;
+        background-color: #90b4ce;
+        color: #fffffe;
+    }
+    /* ボタン色変更ここまで */
+    </style>
+    """
