+++ test.py
@@ -1,0 +1,4 @@
+import random
+
+for i in range(30):
+    print(random.randrange(0, 4))
